Extract the menu to its own init routine
